Redo Database Schema
I was waffling around between candidates being polymorphic; between
whether candidates owned awards or awards owned candidates; what should
be a join table and what shouldn't be; whether a Tally should be
ephemeral or not; and as a result my git history went to crap.

What I finally settled one would be incredible hard to decipher from
this commit but I'll explain it here:

Candidates are records that has a polymorphic source. Eg an Anime can
have multiple candidacies for multiple awards.

An AwardCategory can persist across time. Eg "Anime of the Year" is an
award category that can be won for multiple, different years. An
AwardSeason refers to the block of time that an award can be voted on.
For example, the "Winter 2016 Preliminaries" is an award season, as is
"Year in Review: Anime Awards 2016". Thus, the "awards" table is
essentially a join table between award categories and award seasons.
Eg, an Award could belong to "Anime Awards 2017" award season, and
the award category "Anime of the Year".

A candidate belongs to a specific award, and if selected, has the
"won" column marked as "true".