Fix frontend voting
Previously, voting would occur on all ballots, not on only ballots
under a specific award. This meant you could be voting for multiple
categories and multiple seasons in the same voting session.

This commit adjusts the controller logic so that you are limited to
the scope of one award per voting session. This commit also adjusts
the views to comply with these changes.